requirements.txt
streamlit
google-generativeai
pandas
python-dotenv
fpdf
pillow
duckduckgo-search